Add files via upload
This file is for the main creation/edit/removal, and displaying of contracts/Contractors 
It is paired with 2 other programs for a seamless interaction.
Please be sure to add Tsecure to any PC you put this program on, as it is recommended to have some form of security for the Admin PC.
To make sure this program is ran after successfully logging into Tsecure. Be sure to edit 'startup' and look for the variable to run a program.